+++ src/frontend/configuration/wording/Wording.tsx
@@ -568,6 +568,9 @@ export interface Wording {
   readonly PARAMETRAGE_TITRE: string;
   readonly PARAMETRAGE_AUTORISATIONS_TITRE: string;
   readonly PARAMETRAGE_AIDE_TITRE: string;
+  readonly PARAMETRAGE_MESSAGE_ACCUEIL_TITRE: string;
+  readonly PARAMETRAGE_DEFINIR_MESSAGE_ACCUEIL: string;
+  readonly PARAMETRAGE_MESSAGE_ACCUEIL_HISTORIQUE: string;
   readonly PARAMETRAGE_EJ_TAB: string;
   readonly PARAMETRAGE_ET_MS_TAB: string;
   readonly PARAMETRAGE_ET_SAN_TAB: string;
@@ -639,7 +642,6 @@ export interface Wording {
   readonly LISTE_NON_TROUVÉE: string;
   //Callout page d'acceuil
   readonly NOUVEAU: string;
-  readonly NOUVELLES_FONCTIONNALITÉS_DATE_FIN: string;
 
   //Vigie RH
   readonly INDICATEURS_HELIOS_BLOC_TITLE: string;
@@ -739,6 +741,9 @@ export interface Wording {
   readonly PARAMETRAGE_AIDE_BOUTON_DESCENDRE_RESSOURCE: string;
   readonly PARAMETRAGE_AIDE_BOUTON_MODIFIER_RESSOURCE: string;
   readonly PARAMETRAGE_AIDE_BOUTON_SUPPRIMER_RESSOURCE: string;
+  readonly PARAMETRAGE_MESSAGE_ACCUEIL_NAV_SECTIONS: string;
+  readonly PERIODE_AFFICHAGE: string;
+  readonly MESSAGE: string;
 
   // Aide
   readonly AIDE_TITRE_PAGE: string;